PC calculation for either jump or branch instructions added
Branch instructions

ALU stage receives the branch_offset and the PC for the actual instruction
in order to add them. Then, both the result and the branch_taken signal
are flopped. Mem stage will send this result to the fetch stage.

Jump instructions

ALU stage sends both the jump address and the is_jump condition to the
fetch stage.

Fetch stage

With all that data, it is cappable to compute the next PC.