Instruction memory has now a 10 cycles delay and is separated from the CPU
When a new PC enters the fetch stage for the first time, it rises a memory
request to the instruction memory and the stage is stalled until the
response with the instruction arrives.

When either a branch (taken) or a jump is detected in the ALU stage, it
is needed to ignore the actual requests to the instruction memory. When
that instruction arrives, it is not propagated to the decode stage and
a new request is done with the new PC.