Improved mem stall logic
MEM stage is now setting stall wire to 0 when the data from the
memory arrives instead of in the next cycle. With this approach,
the fetch stage is one cycle less stalled.

The fetch stage FSM has also been improved: all the logic for the
next cycle is ready when the instruction from the memory arrives. If
the mem stage is stalled, the propagation to the decode stage will
not take place, but its logic will be hold.