Reg write enable wire is flopped with 1 only one cycle
For non load instructions, the value is flopped set to 1 next
cycle the instruction reached the memory stage.

For load instructions, the value is flopped set to 1 when the data
from the memory arrives.

For this reason, it is no longer needed to mask it with the valid
in the wb stage.