Non memory instructions are able to write back at ALU stage
Since we don't have a data structure that allows the process to handle
with out of order completions, it is needed that the instruction that
is inside decode stage has to check whether the memory stage is stalled
or not. If this is the case, it also has to stall in order to not allow
the instruction to go to the ALU stage.