Updated logic for stalling both Fetch and Decode stages
Fetch stage is only stalled when Decode stage is so.

Memory stage finishing its stalling phase means that the following
cycle the instruction will be at the write back stage, so it is
not possible to propagate the instruction in the Decode stage to
the ALU stage, same will happen with executions tage.
A wire indicating Decode stage that in the next cycle a write back
is going to happen has been added in order to stall the stage an
additional cycle.