Decode-Execution stages RAW hazards implemented
If there is an instruction that needs to read operands in the decode
stage, and one of those are the destination register of a multiplication
instruction in any of the dxecution stages, the instruction must get
stalled.

For this RAW hazard, it is only needed to check for EX1-EX4 stages,
since the result for a multiplication is valid in EX5 stage. Therefore,
a bypass can be used to decode stage.