Not allow memory instructions to overtake multiplication in early stages
Taking into account that a memory operation takes at least 2 cycles
in the memory stage (it hits on cache), we need to stall that
instruction in the decode stage either if there is a multiplication
in EX1 or EX2. Otherwise we will end up in the situations either the
memory instruction is in the write back stage when the multiplication
is still in EX5 stage or the memory instruction is in the last cycle
of the memory stage and the multiplication is in EX5 (they both access
to the write back stage next cycle).