WB stage changed to WB arbiter
WB arbiter is the responsible to choose which is the instruction
that is able to complete among the instructions in the ALU, MEM or
EX5 stages. The priority is: MEM > EX5 > ALU.

When an instruction is not granted to complete, it has to remain in
its current stage until is able to do so.

Destination register is written at the end of the cycle of an
instruction compliting, so the stable result will be visible in the
next cycle