New ALU bubble condition added
Since out-of-order completions are not allowed yet, we must not allow
an instruction in decode stage to advance to ALU stage if the current
instruction in the ALU stage is a memory instruction.
With this, we prevent an out-of-order completion in the next cycle
while the memory operation is serving a miss.